Aiming and reloading
I have now added in the Aiming in and set the the IK(Inverse Kinematics) for the hands for the player. So this now positions both left hand and right to hold the weapon correctly. This also allows the weapon to aim accurately based one where the player(Camera) is pointing. I have also added the reload animations and set up the StateMachine for the Animator for the player to know when the reload animation is complete and when to set _reload to false.